All header-only dependencies are now submodules
Easier setup since you now don't have to build boost and cmake anymore and add their path to the system variables
