+++ .gitmodules
@@ -1,13 +1,48 @@
 [submodule "external/hnswlib"]
 	path = external/hnswlib
 	url = https://github.com/nmslib/hnswlib.git
+	branch = master
 [submodule "external/spdlog"]
 	path = external/spdlog
 	url = https://github.com/gabime/spdlog
+	branch = v1.x
 [submodule "external/glfw"]
 	path = external/glfw
 	url = https://github.com/glfw/glfw/
+	branch = 3.3-stable
 [submodule "python-wrapper/pybind11"]
 	path = python_wrapper/pybind11
 	url = https://github.com/pybind/pybind11
 	branch = stable
+[submodule "external/eigen"]
+	path = external/eigen
+	url = https://gitlab.com/libeigen/eigen.git
+	branch = master
+[submodule "external/boost/histogram"]
+	path = external/boost/histogram
+	url = https://github.com/boostorg/histogram
+	branch = master
+[submodule "external/boost/core"]
+	path = external/boost/core
+	url = https://github.com/boostorg/core
+	branch = master
+[submodule "external/boost/mp11"]
+	path = external/boost/mp11
+	url = https://github.com/boostorg/mp11
+	branch = master
+[submodule "external/boost/assert"]
+	path = external/boost/assert
+	url = https://github.com/boostorg/assert
+	branch = master
+[submodule "external/boost/throw_exception"]
+	path = external/boost/throw_exception
+	url = https://github.com/boostorg/throw_exception
+	branch = master
+[submodule "external/boost/config"]
+	path = external/boost/config
+	url = https://github.com/boostorg/config
+	branch = master
+[submodule "external/boost/variant2"]
+	path = external/boost/variant2
+	url = https://github.com/boostorg/variant2
+	branch = master
